requirements.txt
# Data manipulation
pandas==2.0.0
numpy==1.24.0

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.0

# Time Series Analysis
statsmodels==0.14.0
prophet==1.1.4

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
pyyaml==6.0
tqdm==4.65.0
joblib==1.3.1

# Development
pytest==7.3.1
black==23.3.0
flake8==6.0.0
jupyter==1.0.0